Use custom trait instead of std::ops for UGen math
The commit introduces an `Input` trait that replaces the `IntoValue`
trait. The `Input` trait implements SuperCollider's binary and unary
operators such as `mul`, `add`, and `midicps. This provides a more
ergonomic interface than implementing the std operators like
`std::ops::Mul`. Both sides of the implementation of a binary operator
cannot be specified as `Into<Value>` because multiple numeric types are
`Into<Value>` and the implementation of the operator would conflict with
simply adding two numbers.